requirements.txt
# GLR Pipeline Dependencies
streamlit>=1.28.0
python-docx>=0.8.11
PyMuPDF>=1.23.0
requests>=2.31.0
python-dotenv>=1.0.0
